chore: Update example applications
