+++ examples/example-fastify-api/.env.example
@@ -1,0 +1,2 @@
+AUTH0_DOMAIN=
+AUTH0_AUDIENCE=
