+++ examples/example-fastify-api/.npmrc
