+++ examples/example-fastify-api/package.json
@@ -1,0 +1,19 @@
+{
+    "name": "example-fastify-api",
+    "version": "1.0.0",
+    "description": "",
+    "type": "module",
+    "scripts": {
+        "start": "tsx src/index.ts --project tsconfig.json"
+    },
+    "devDependencies": {
+        "ts-node": "^10.9.2",
+        "tsx": "^4.19.2",
+        "typescript": "^5.7.3"
+    },
+    "dependencies": {
+        "@auth0/auth0-fastify": "*",
+        "dotenv": "^16.4.7",
+        "fastify": "^5.2.1"
+    }
+}
