+++ examples/example-fastify-web/.env.example
@@ -1,0 +1,5 @@
+AUTH0_DOMAIN=
+AUTH0_CLIENT_ID=
+AUTH0_CLIENT_SECRET=
+AUTH0_SESSION_SECRET=
+APP_BASE_URL=http://localhost:3000
