+++ examples/example-fastify-web/.npmrc
@@ -1,0 +1,1 @@
+registry=https://registry.npmjs.org
