+++ examples/example-fastify-web/package.json
@@ -1,10 +1,11 @@
 {
-    "name": "example-fastify",
+    "name": "example-fastify-web",
     "version": "1.0.0",
     "description": "",
     "type": "module",
     "scripts": {
-        "start": "tsx src/index.ts --project tsconfig.json"
+        "start": "tsx src/index.ts --project tsconfig.json",
+        "build": "tsc --project tsconfig.json"
     },
     "devDependencies": {
         "@types/ejs": "^3.1.5",
@@ -13,7 +14,9 @@
         "typescript": "^5.7.3"
     },
     "dependencies": {
+        "@fastify/static": "^8.1.1",
         "@fastify/view": "^10.0.2",
+        "dotenv": "^16.4.7",
         "ejs": "^3.1.10",
         "fastify": "^5.2.1"
     }
