+++ examples/example-fastify-web/views/layout.ejs
@@ -1,0 +1,63 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>Auth0-Fastify Web Application demo</title>
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+      integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH"
+      crossorigin="anonymous"
+    />
+  </head>
+  <body>
+    <nav class="py-2 bg-body-tertiary border-bottom">
+      <div class="container d-flex flex-wrap">
+        <a
+          href="/"
+          class="d-flex align-items-center mb-2 mb-lg-0 text-white text-decoration-none"
+        >
+          <img src="/img/auth0.png" width="36" height="36" alt="Auth0 logo" />
+        </a>
+        <ul class="nav me-auto">
+          <li class="nav-item">
+            <a
+              href="/public"
+              class="nav-link link-body-emphasis px-2 active"
+              aria-current="page"
+              >Public Page</a
+            >
+          </li>
+          <li class="nav-item">
+            <a
+              href="/private"
+              class="nav-link link-body-emphasis px-2 active"
+              aria-current="page"
+              >Private Page</a
+            >
+          </li>
+        </ul>
+        <ul class="nav">
+          <li class="nav-item">
+            <% if(isLoggedIn){ %>
+            <a class="nav-link link-body-emphasis px-2" href="/auth/logout"
+              >Log out <%- locals.user ? '(' + locals.user.name + ')' : '' %></a
+            >
+            <% } else{ %>
+            <a class="nav-link link-body-emphasis px-2" href="/auth/login"
+              >Log in</a
+            >
+            <% } %>
+          </li>
+        </ul>
+      </div>
+    </nav>
+    <div class="container py-4"><%- body %></div>
+    <script
+      src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
+      integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
+      crossorigin="anonymous"
+    ></script>
+  </body>
+</html>
