+++ examples/example-fastify-web/views/private.ejs
@@ -1,0 +1,1 @@
+This is a private page.
