+++ examples/example-fastify-web/views/public.ejs
@@ -1,0 +1,1 @@
+This is a public page.
