--- examples/example-fastify/views/index.ejs
@@ -1,13 +1,0 @@
-<!-- index.ejs --->
-<!DOCTYPE html>
-<html lang="en">
-  <head></head>
-  <body>
-    <% if(isLoggedIn){ %>
-      <h1>Hello, <%= name %>!</h1>
-      <a href="/auth/logout">Log out</a>
-    <% } else{ %>  
-      <a href="/auth/login">Log in</a>
-   <% } %>
-  </body>
-</html>
