--- examples/example-fastify/views/profile.ejs
@@ -1,9 +1,0 @@
-<!-- index.ejs --->
-<!DOCTYPE html>
-<html lang="en">
-  <head></head>
-  <body>
-    <h1>Hello, <%= name %>!</h1>
-    <a href="/auth/logout">Log out</a>
-  </body>
-</html>
