+++ package.json
@@ -2,7 +2,8 @@
   "name": "auth0-server-js",
   "version": "1.0.0",
   "workspaces": [
-    "packages/*"
+    "packages/*",
+    "examples/*"
   ],
   "scripts": {
     "build": "turbo run build",
