+++ packages/auth0-fastify/package.json
@@ -24,8 +24,6 @@
     }
   },
   "devDependencies": {
-    "@auth0/auth0-api-js": "*",
-    "@auth0/auth0-server-js": "*",
     "@eslint/js": "^9.20.0",
     "@vitest/coverage-v8": "^3.0.6",
     "eslint": "^9.20.1",
@@ -35,6 +33,8 @@
     "vitest": "^3.0.5"
   },
   "dependencies": {
+    "@auth0/auth0-api-js": "*",
+    "@auth0/auth0-server-js": "*",
     "@fastify/cookie": "^11.0.2",
     "fastify": "^5.2.1",
     "fastify-plugin": "^5.0.1"
